requirements.txt
absl-py==0.7.1
astor==0.8.0
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
Markdown==3.1.1
mock==3.0.5
numpy==1.16.4
pandas==0.24.2
protobuf==3.8.0
python-dateutil==2.8.0
python-mnist==0.6
pytz==2019.1
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Werkzeug==0.15.4
django==2.2.2
djangorestframework==3.9.4